// src/app/core/models/dossier.ts
export class Dossier {
  id: string;
  nom_cocontractant: any;
  lieu_execution: any;
  montant: any;
  date_signature: any;
  duree: any;
  date_fin: any;
  type: any;
  sous_typologie: any;
  reference: any;
  piece_jointe: any;
  date_alerte: any;
  emails: any;
  client: any;
  intitule: any;
  objetEmail: any;
  dateNotification: any;
  datePremierNotif: any;
  dateSecondNotif: any;
  etatNotif: any;
}